[ASL-4523] detect incognito
