--- packages/core/src/constants/domId.ts
@@ -1,1 +1,0 @@
-export const LEAVE_IAB_MODAL_ID = 'qubic-connect-leave-iab-modal';
