+++ packages/core/src/constants/index.ts
@@ -1,4 +1,3 @@
 export { API_URL, AUTH_REDIRECT_URL, MARKET_API_URL } from './backend';
-export { LEAVE_IAB_MODAL_ID } from './domId';
 export { PaymentErrorCode } from './errorCodes';
 export { EXTERNAL_PROVIDER_MAP } from './externalProvider';
