+++ packages/detect-iab/package.json
@@ -31,6 +31,7 @@
     "@ethersproject/providers": "^5.6.8",
     "@qubic-js/detect-inapp": "^1.4.1",
     "clsx": "^1.1.1",
+    "detectincognitojs": "^1.3.0",
     "jss": "^10.9.0",
     "jss-preset-default": "^10.9.0",
     "preact": "^10.8.2"
