+++ packages/detect-iab/src/LeaveInAppBrowserModal.tsx
@@ -1,5 +1,3 @@
-import jss from 'jss';
-import preset from 'jss-preset-default';
 import { useEffect, memo, useCallback, useState } from 'preact/compat';
 import clsx from 'clsx';
 import InApp, { InAppBrowser } from '@qubic-js/detect-inapp';
@@ -7,163 +5,13 @@ import InApp, { InAppBrowser } from '@qubic-js/detect-inapp';
 import HorizontalMoreOptions from './icons/HorizontalMoreOptions';
 import VerticalMoreOptions from './icons/VerticalMoreOptions';
 import ExportMoreOptions from './icons/ExportMoreOptions';
-import TooltipArrow from './icons/TooltipArrow';
-import { LEAVE_IAB_MODAL_ID, doubleBrowserIcons } from './constant';
+import { LEAVE_IAB_MODAL_ID } from './constant';
 import { getLocaleStrings } from './locale';
-
-// One time setup with default plugins and settings.
-jss.setup(preset());
+import { classes } from './styles';
+import Hint from './Hint';
+import { doubleBrowserIcons } from './icons/DoubleBrowserIcons';
 
 const localeStrings = getLocaleStrings();
-
-const { classes } = jss
-  .createStyleSheet({
-    modal: {
-      position: 'fixed',
-      top: 0,
-      left: 0,
-      width: '100vw',
-      height: '100vh',
-      alignItems: 'center',
-      justifyContent: 'center',
-      display: 'flex',
-      zIndex: 100,
-      touchAction: 'none',
-    },
-    backdrop: {
-      zIndex: -1,
-      backgroundColor: 'rgba(0, 0, 0, 0.7)',
-      position: 'fixed',
-      width: '100vw',
-      height: '100vh',
-      display: 'flex',
-      flexDirection: 'column',
-      justifyContent: 'center',
-      alignItems: 'center',
-      backdropFilter: 'saturate(180%) blur(10px)',
-      WebkitBackdropFilter: 'saturate(180%) blur(10px)',
-    },
-    contentWrapper: {
-      padding: '0 24px',
-      display: 'flex',
-      flexDirection: 'column',
-      justifyContent: 'flex-start',
-      alignItems: 'center',
-    },
-    emoji: {
-      fontSize: 60,
-      lineHeight: '60px',
-    },
-    alertSentence1: {
-      fontSize: 16,
-      lineHeight: '24px',
-      color: '#FFF',
-      fontWeight: 600,
-      textAlign: 'center',
-      marginTop: 16,
-    },
-    alertSentence2: {
-      fontSize: 14,
-      lineHeight: '18px',
-      color: '#FFF',
-      fontWeight: 400,
-      textAlign: 'center',
-      marginTop: 8,
-    },
-    hintWrapper: {
-      maxWidth: 'calc(100% - 16px)',
-      display: 'flex',
-      flexDirection: 'column',
-      justifyContent: 'flex-start',
-      alignItems: 'flex-start',
-      right: 8,
-    },
-    hintWrapperTop: {
-      position: 'absolute',
-      top: 20,
-    },
-    hintWrapperBottom: {
-      position: 'absolute',
-      bottom: 20,
-    },
-    dialogWrapper: {
-      borderRadius: 8,
-      padding: '8px 16px',
-      backgroundColor: '#2962FF',
-      position: 'relative',
-    },
-    dialogText: {
-      fontFamily: 'Roboto, sans-serif',
-      fontSize: 16,
-      lineHeight: '24px',
-      color: '#FFF',
-      fontWeight: 400,
-      display: 'flex',
-      flexDirection: 'row',
-      justifyContent: 'flex-start',
-      alignItems: 'center',
-      margin: 0,
-      flexWrap: 'wrap',
-    },
-    icon: {
-      width: 24,
-      height: 24,
-    },
-    exportMoreIcon: {
-      width: 16,
-      height: 16,
-      paddingLeft: 8,
-      paddingRight: 8,
-    },
-    arrowTopWrapper: {
-      display: 'flex',
-      position: 'absolute',
-      bottom: 'calc(100% - 1px)',
-      right: 12,
-    },
-    arrowBottomWrapper: {
-      display: 'flex',
-      position: 'absolute',
-      top: 'calc(100% - 1px)',
-      right: 12,
-      transform: 'rotate(180deg) scaleX(-1)',
-    },
-    currentUrlWrapper: {
-      display: 'flex',
-      flexDirection: 'row',
-      borderRadius: 8,
-      width: 'calc(100% - 48px)',
-      backgroundColor: 'white',
-      padding: 4,
-      marginTop: 24,
-    },
-    currentUrlInput: {
-      height: 40,
-      flex: 1,
-      minWidth: 0,
-      border: 'none',
-      userSelect: 'all',
-      borderRadius: 8,
-      fontSize: 16,
-      lineHeight: '24px',
-      '&:focus': {
-        outline: 'none',
-      },
-    },
-    currentUrlBtn: {
-      flexShrink: 0,
-      border: 0,
-      color: '#2962FF',
-      fontSize: 16,
-      lineHeight: '24px',
-      backgroundColor: 'transparent',
-      padding: '8px 16px',
-    },
-    currentUrlBtnCopied: {
-      color: '#5FC417',
-    },
-  })
-  .attach();
 
 export interface ShowBlockerOptions {
   redirectUrl?: string;
@@ -315,16 +163,6 @@ const LeaveInAppBrowserModal = memo<ModalProps>(props => {
 
   const arrowPosition = getArrowPosition(platform, currentBrowser);
 
-  const hintWrapperClass = (() => {
-    switch (arrowPosition) {
-      case 'top':
-        return clsx(classes.hintWrapper, classes.hintWrapperTop);
-      case 'bottom':
-      default:
-        return clsx(classes.hintWrapper, classes.hintWrapperBottom);
-    }
-  })();
-
   useEffect(() => {
     const body = document.querySelector('body');
     if (body) {
@@ -379,8 +217,8 @@ const LeaveInAppBrowserModal = memo<ModalProps>(props => {
       <div className={classes.backdrop}>
         <div className={classes.contentWrapper}>
           <img src={`data:image/png;base64, ${doubleBrowserIcons}`} alt="browser" style={{ width: 120, height: 120 }} />
-          <span className={classes.alertSentence1}>{localeStrings.alertSentence1}</span>
-          <span className={classes.alertSentence2}>{localeStrings.alertSentence2}</span>
+          <span className={classes.alertSentence1}>{localeStrings.alertIabSentence1}</span>
+          <span className={classes.alertSentence2}>{localeStrings.alertIabSentence2}</span>
           {displayMode === DisplayMode.COPY && (
             <div className={classes.currentUrlWrapper}>
               <input readOnly value={redirectUrl} className={classes.currentUrlInput} />
@@ -398,20 +236,13 @@ const LeaveInAppBrowserModal = memo<ModalProps>(props => {
         </div>
       </div>
       {displayMode === DisplayMode.HINT && (
-        <div className={hintWrapperClass}>
-          <div className={classes.dialogWrapper}>
-            <p className={classes.dialogText}>
-              {localeStrings.hintPart1}
-              {getHintImage(platform, currentBrowser)}
-              {localeStrings.hintPart2}
-              {getHintBrowserString(platform, currentBrowser)}
-              {localeStrings.hintPart3}
-            </p>
-            <div className={arrowPosition === 'top' ? classes.arrowTopWrapper : classes.arrowBottomWrapper}>
-              <TooltipArrow />
-            </div>
-          </div>
-        </div>
+        <Hint arrowPosition={arrowPosition}>
+          {localeStrings.hintPart1}
+          {getHintImage(platform, currentBrowser)}
+          {localeStrings.hintPart2}
+          {getHintBrowserString(platform, currentBrowser)}
+          {localeStrings.hintPart3}
+        </Hint>
       )}
     </div>
   );
