+++ packages/detect-iab/src/locale.ts
@@ -1,21 +1,26 @@
 const en = {
-  alertSentence1: 'You are using a third-party browser',
-  alertSentence2: 'To have a better experience, open the website in system browser.',
+  alertIabSentence1: 'You are using a third-party browser',
+  alertIabSentence2: 'To have a better experience, open the website in system browser.',
+  alertIncognitoSentence1: 'You are now in private browsing mode',
+  alertIncognitoSentence2: 'To have a better experience, open the website in normal browsing mode.',
   hintPart1: 'Tap "',
   hintPart2: '" and choose "',
   hintPart3: '".',
   openInBrowser: 'Open in browser',
   openInSystemBrowser: 'Open in system browser',
   openInDefaultBrowser: 'Open in your default browser',
-  openInSafari: 'Open in safari',
-  openInChrome: 'Open in chrome',
+  openInSafari: 'Open in Safari',
+  openInSafariInNormal: 'Open in Safari in normal mode',
+  openInChrome: 'Open in Chrome',
   copy: 'Copy',
   copied: 'Copied',
 };
 
 const zhTW = {
-  alertSentence1: '您正在使用第三方瀏覽器',
-  alertSentence2: '為了享有更完整的服務，請以系統瀏覽器開啟網頁',
+  alertIabSentence1: '您正在使用第三方瀏覽器',
+  alertIabSentence2: '為了享有更完整的服務，請以系統瀏覽器開啟網頁',
+  alertIncognitoSentence1: '您正在使用私密瀏覽模式',
+  alertIncognitoSentence2: '為了享有更完整的服務，請在瀏覽器以一般模式開啟網頁',
   hintPart1: '點擊「',
   hintPart2: '」，並選擇「',
   hintPart3: '」',
@@ -23,6 +28,7 @@ const zhTW = {
   openInSystemBrowser: '以瀏覽器開啟',
   openInDefaultBrowser: '使用預設瀏覽器開啟',
   openInSafari: '在 Safari 開啟',
+  openInSafariInNormal: '在 Safari 以一般模式開啟',
   openInChrome: '在 Chrome 開啟',
   copy: '複製',
   copied: '已複製',
